+++ tests/js/highway_chord_render_cache.test.js
@@ -19,8 +19,23 @@ const path = require('node:path');
 
 const highwayJs = path.join(__dirname, '..', '..', 'static', 'highway.js');
 
+
+// R3c: highway.js is being carved into modules, so its source is no longer ONE file. Read the
+// whole set rather than re-pinning at whichever file currently holds a function — re-pinning
+// just breaks again next time, and a source-shape assertion that silently stops finding its
+// target is indistinguishable from one that passes.
+function highwaySources() {
+    const root = path.join(__dirname, '..', '..');
+    const jsDir = path.join(root, 'static', 'js');
+    const parts = [fs.readFileSync(path.join(root, 'static', 'highway.js'), 'utf8')];
+    for (const f of fs.readdirSync(jsDir).sort()) {
+        if (f.startsWith('highway-') && f.endsWith('.js')) parts.push(fs.readFileSync(path.join(jsDir, f), 'utf8'));
+    }
+    return parts.join('\n');
+}
+
 test('_ensureChordRenderCache keys off src, _inverted, AND chordTemplates', () => {
-    const src = fs.readFileSync(highwayJs, 'utf8');
+    const src = highwaySources();
     // The cache key triple must include chordTemplates — without it, a
     // late-arriving `chord_templates` WS message leaves cached
     // nonZeroNotes / nonZeroFrets stale until the next chord transition.
@@ -41,7 +56,7 @@ test('_ensureChordRenderCache keys off src, _inverted, AND chordTemplates', () =
 });
 
 test('chordTemplates change resets fretline preview and frame-mismatch warner', () => {
-    const src = fs.readFileSync(highwayJs, 'utf8');
+    const src = highwaySources();
     // The cache-invalidation block must clear both _chordFretLineNotes
     // (so _updateFretLinePreview re-publishes with corrected isOpen
     // classification) and _frameMismatchWarned (so a chord ID warned
